+++ src/pages/app/components/index.ts
@@ -2,3 +2,5 @@ export * from "./completion";
 export * from "./speech";
 export * from "./speech/audio-visualizer";
 export * from "./speech/StatusIndicator";
+export * from "./TranscriptionPanel";
+export * from "./AIResponsePanel";